test: clean up hir tests
* Factor out reusable components to be used for unit tests across
  various HIR crates
* Make use of recent ergonomics improvments, e.g. `ProgramBuilder`
